+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+ * SVG exporter improvements
 
 ## 0.6.2 - 2020-04-30
  * Support for all importers / exporters in lottie GUI